Refactor names and packages
