+++ app/src/main/AndroidManifest.xml
@@ -56,11 +56,11 @@
             tools:ignore="LockedOrientationActivity" />
 
         <activity
-            android:name=".fragment_case.VGSCollectFragmentActivity"
+            android:name=".collect_fragment.VGSCollectFragmentActivity"
             android:label="Collect (Fragment)" />
 
         <activity
-            android:name=".viewpager_case.VGSViewPagerActivity"
+            android:name=".collect_viewpager.ViewPagerActivity"
             android:label="Collect (ViewPager)" />
 
         <activity
@@ -76,11 +76,11 @@
             android:label="Instrumented Demo" />
 
         <activity
-            android:name=".compose.activity_case.ComposeActivity"
+            android:name=".collect_compose.activity_case.ComposeActivity"
             android:label="Compose Demo (Activity)" />
 
         <activity
-            android:name=".compose.fragment_case.ComposeFragmentActivity"
+            android:name=".collect_compose.fragment_case.ComposeFragmentActivity"
             android:label="Compose Demo (Fragment)" />
 
         <meta-data
