+++ app/src/main/res/layout/activity_fragment_compose.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <androidx.fragment.app.FragmentContainerView xmlns:android="http://schemas.android.com/apk/res/android"
     android:id="@+id/fragment"
-    android:name="com.verygoodsecurity.demoapp.compose.fragment_case.ComposeFragment"
+    android:name="com.verygoodsecurity.demoapp.collect_compose.fragment_case.ComposeFragment"
     android:layout_width="match_parent"
     android:layout_height="match_parent" />
